pyproject.toml
[project]
name = "sispubli-api"
version = "1.1.0"
description = "API REST para extracao de certificados do sistema Sispubli/IFS"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "cryptography>=46.0.7",
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "pydantic>=2.13.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "uvicorn>=0.44.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "pytest-mock>=3.15.1",
    "pytest-recording>=0.13.4",
    "ruff>=0.15.10",
    "vcrpy>=8.1.1",
]

# =============================================================================
# Ruff — Linter + Formatter (substitui flake8 + black)
# =============================================================================

[tool.ruff]
target-version = "py313"
line-length = 100
extend-exclude = ["*.html", "*.har", "**/resources/*"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort (ordenacao de imports)
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "S",    # flake8-bandit (seguranca)
    "T201", # flake8-print (proibe prints no codigo)
]
ignore = [
    "S101",  # assert em testes e OK
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "S105", "S104", "E501"]  # Permitir asserts, tokens e linhas longas em testes

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# =============================================================================
# Pytest — Configuracao de testes
# =============================================================================

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-m 'not e2e' --cov=src --cov=. --cov-report=term-missing --cov-report=html --tb=short"
markers = [
    "e2e: Testes end-to-end que acessam o Sispubli real (localmente via pre-push)",
]

# =============================================================================
# Coverage — Configuracao do pytest-cov
# =============================================================================

[tool.coverage.run]
omit = [
    "tests/*",
    ".venv/*",
    "__pycache__/*",
    "main.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.",
    "pass",
]
show_missing = true
